Add files via upload
•  A recent startup started by the 3 friends from IIT Delhi. The company deals with apparel and trending millennium clothing and accessories. 
Problem Statement
The company wants to expand across India through an online web market and various offline stores. Still, before taking this step, they wanted to be double-sure regarding the demand and the sector they should target. The company has a limited cash reserve and they want to utilize it in the best possible way.